Fix nodejs stack - nss_wrapper command was uninstalling node
The default nss_wrapper.install script installs and uninstalls
make. The uninstall step removes node and causes the resulting
image to be useless.

Images inheriting from FROM registry.centos.org/che-stacks/centos-nodejs
come with nodejs already installed, so the command has to be modified
slightly. This is not a problem in other dockerfiles, as we manually
install nodejs at a later step.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>