use cmake3 in nss_wrapper.install (fixes #50)
This requires using EPEL (core CentOS only has cmake 2.x).

An alternative solution would be to to not build nss_wrapper from
its (master!) source, but instead to install it from a package;
but that also requires EPEL, and would likely lead to an older
(but stable..) version of nss_wrapper than build from source.

Full background story and analysis is documented in:
https://github.com/redhat-developer/che-dockerfiles/issues/50

Signed-off-by: Michael Vorburger <mike@vorburger.ch>